requirements.txt
Pillow
matplotlib
odrive